requirements.txt
openai
chromadb
langchain
pypdf
tiktoken
requests
python-docx
docx2txt
